requirements.txt
# Core dependencies
openai>=1.12.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tiktoken>=0.5.2

# Testing
pytest>=8.0.0
httpx>=0.26.0
